# services/iam/doc/openapi.json
{
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ],
  "openapi": "3.0.0",
  "info": {
    "description": "This is a the API documentation for the OHI IAM system",
    "version": "1.0.0",
    "title": "IAM Api for OIH",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "info@openintegrationhub.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/login": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login route",
        "description": "",
        "operationId": "performLogin",
        "requestBody": {
          "description": "Login with credentials",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "400": {
            "description": "Bad request; Missing mandatory body"
          },
          "403": {
            "description": "Cannot login user. Account is either disabled or a precondition is missing."
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Register route",
        "description": "",
        "operationId": "performRegister",
        "requestBody": {
          "description": "Register new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered account successfully"
          },
          "400": {
            "description": "Bad request, Missing mandatory body"
          }
        }
      }
    },
    "/logout": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Logout route",
        "description": "",
        "operationId": "preformLogout",
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "401": {
            "description": "User is not logged in."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Login": {
        "title": "Login data",
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "Register": {
        "title": "Register data",
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "tenant": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "firstname",
          "lastname",
          "tenant"
        ]
      },
...
      "AccessToken": {
        "title": "Access token object",
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Auth token"
          },
          "id": {
            "type": "string",
            "description": "Token id"
          }
        }
      },
...
    }
  }
}